+++ internal/processing/helpers.go
@@ -1,11 +1,14 @@
 package processing
 
 import (
+	"context"
 	"fmt"
+
 	gatewayv1alpha1 "github.com/kyma-incubator/api-gateway/api/v1alpha1"
 	gatewayv1beta1 "github.com/kyma-incubator/api-gateway/api/v1beta1"
 	"github.com/kyma-incubator/api-gateway/internal/builders"
 	k8sMeta "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
 var (
@@ -41,4 +44,54 @@ func GetOwnerLabels(api *gatewayv1beta1.APIRule) map[string]string {
 	labels := make(map[string]string)
 	labels[OwnerLabelv1alpha1] = fmt.Sprintf("%s.%s", api.ObjectMeta.Name, api.ObjectMeta.Namespace)
 	return labels
+}
+
+func Evaluate(ctx context.Context, client client.Client, apiRule *gatewayv1beta1.APIRule, processors []ReconciliationProcessor) error {
+	for _, processor := range processors {
+		objectChanges, err := processor.EvaluateReconciliation(ctx, client, apiRule)
+		if err != nil {
+			return err
+		}
+
+		err = applyChanges(ctx, client, objectChanges...)
+		if err != nil {
+			//  "We don't know exactly which object(s) are not updated properly. The safest approach is to assume nothing is correct and just use `StatusError`."
+			return err
+		}
+	}
+	return nil
+}
+
+// applyChanges applies the given commands on the cluster
+func applyChanges(ctx context.Context, client client.Client, changes ...*ObjectChange) error {
+
+	for _, change := range changes {
+		err := applyChange(ctx, client, change)
+		if err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+func applyChange(ctx context.Context, client client.Client, change *ObjectChange) error {
+	var err error
+
+	switch change.Action {
+	case "create":
+		err = client.Create(ctx, change.Obj)
+	case "update":
+		err = client.Update(ctx, change.Obj)
+	case "delete":
+		err = client.Delete(ctx, change.Obj)
+	default:
+		err = fmt.Errorf("apply action %s is not supported", change.Action)
+	}
+
+	if err != nil {
+		return err
+	}
+
+	return nil
 }
